+++ torchmetrics/regression/r2score.py
@@ -16,7 +16,7 @@ from typing import Any, Callable, Optional
 import torch
 from torch import Tensor, tensor
 
-from torchmetrics.functional.regression.r2score import _r2score_compute, _r2score_update
+from torchmetrics.functional.regression.r2 import _r2_score_compute, _r2_score_update
 from torchmetrics.metric import Metric
 
 
@@ -135,7 +135,7 @@ class R2Score(Metric):
             preds: Predictions from model
             target: Ground truth values
         """
-        sum_squared_error, sum_error, residual, total = _r2score_update(preds, target)
+        sum_squared_error, sum_error, residual, total = _r2_score_update(preds, target)
 
         self.sum_squared_error += sum_squared_error
         self.sum_error += sum_error
@@ -144,7 +144,7 @@ class R2Score(Metric):
 
     def compute(self) -> Tensor:
         """Computes r2 score over the metric states."""
-        return _r2score_compute(
+        return _r2_score_compute(
             self.sum_squared_error, self.sum_error, self.residual, self.total, self.adjusted, self.multioutput
         )